Refactor entrypoint script to improve AWS and GCP deployment handling, including credential management and repository URI definitions.

+++ build-push-image/entrypoint.sh
@@ -12,9 +12,24 @@ SHORT_SHA=$(git rev-parse --short=7 HEAD)
 # Build and Push Docker image
 REPOSITORY_NAME=$(basename "$GITHUB_REPOSITORY")
 
+if [$DEPLOY_PROVIDER== "GCP"]; then
+    SERVICE_ACCOUNT_KEY=$GCP_SERVICE_ACCOUNT_KEY
+elif [$DEPLOY_PROVIDER == "AWS"]; 
+    $AWS_CREDS=$AWS_SERVICE_ACCOUNT_KEY
+    eval $AWS_CREDS
+else
+    echo "DEPLOY_PROVIDER não definido ou inválido."
+    exit 1
+fi
 
+elif [$DEPLOY_PROVIDER == "AWS"]; then
+    echo "DEPLOY_PROVIDER is set to AWS"
+else
+    echo "DEPLOY_PROVIDER is not set or invalid."
+    exit 1
+fi
 
-if DEPLOY_PROVIDER == "GCP"
+if [$DEPLOY_PROVIDER == "GCP"];  then
     # Definir REPOSITORY_URI para a branch
     case "$GITHUB_REF_NAME" in "staging" )
       REPOSITORY_URI_BRANCH="us-docker.pkg.dev/image-registry-326015/$REPOSITORY_NAME/staging" ;;
@@ -38,25 +53,27 @@ if DEPLOY_PROVIDER == "GCP"
       echo "Branch not supported: $GITHUB_REF_NAME"
       exit 1
     esac    
-elif DEPLOY_PROVIDER == "AWS"
-  REPOSITORY_URI_BRANCH = 
+elif [$DEPLOY_PROVIDER == "AWS"]; then
+    REPOSITORY_URI_BRANCH_HML="756376728940.dkr.ecr.sa-east-1.amazonaws.com/$REPOSITORY_NAME"  
+    REPOSITORY_URI_BRANCH_PRD="715663453372.dkr.ecr.sa-east-1.amazonaws.com/$REPOSITORY_NAME"
+
     case "$GITHUB_REF_NAME" in "staging" )
-      REPOSITORY_URI_BRANCH="756376728940.dkr.ecr.sa-east-1.amazonaws.com/$REPOSITORY_NAME" ;;
+      REPOSITORY_URI_BRANCH=$REPOSITORY_URI_BRANCH_HML ;;
     
     "master")  
-      REPOSITORY_URI_BRANCH="715663453372.dkr.ecr.sa-east-1.amazonaws.com/$REPOSITORY_NAME:master" ;;
+      REPOSITORY_URI_BRANCH=$REPOSITORY_URI_BRANCH_HML ;;
     
     "main")  
-      REPOSITORY_URI_BRANCH="715663453372.dkr.ecr.sa-east-1.amazonaws.com/$REPOSITORY_NAME:master" ;;
+      REPOSITORY_URI_BRANCH=$REPOSITORY_URI_BRANCH_HML;;
     
     "develop")
-      REPOSITORY_URI_BRANCH="756376728940.dkr.ecr.sa-east-1.amazonaws.com/$REPOSITORY_NAME:develop";;
+      REPOSITORY_URI_BRANCH=$REPOSITORY_URI_BRANCH_HML;;
     
     "release")
-      REPOSITORY_URI_BRANCH="715663453372.dkr.ecr.sa-east-1.amazonaws.com/$REPOSITORY_NAME:release";;
+      REPOSITORY_URI_BRANCH=$REPOSITORY_URI_BRANCH_HML;;
     
     "homolog")
-      REPOSITORY_URI_BRANCH="756376728940.dkr.ecr.sa-east-1.amazonaws.com/$REPOSITORY_NAME:homolog";;
+      REPOSITORY_URI_BRANCH=$REPOSITORY_URI_BRANCH_HML;;
     * )
       
       echo "Branch not supported: $GITHUB_REF_NAME"
@@ -68,24 +85,40 @@ else
 
 fi
 # Definir REPOSITORY_URI para master
-REPOSITORY_URI_PRD="us-docker.pkg.dev/image-registry-326015/$REPOSITORY_NAME/master"
+if [$DEPLOY_PROVIDER == "GCP"]; then
+    REPOSITORY_URI_PRD="us-docker.pkg.dev/image-registry-326015/$REPOSITORY_NAME/master"
+elif [$DEPLOY_PROVIDER == "AWS"]; then
+    REPOSITORY_URI_PRD=$REPOSITORY_URI_BRANCH_PRD
+fi
 
 echo "REPOSITORY_URI_BRANCH: $REPOSITORY_URI_BRANCH"
 echo "REPOSITORY_URI_PRD: $REPOSITORY_URI_PRD"
 
+
+
 # Validar se a secret está em Base64
-if echo "$GCP_SERVICE_ACCOUNT_KEY" | base64 -d &>/dev/null; then
+if echo "$SERVICE_ACCOUNT_KEY" | base64 -d &>/dev/null; then
     echo "Decodificando secret em Base64..."
-    echo "$GCP_SERVICE_ACCOUNT_KEY" | base64 -d > gcp-sa.json
+    echo "$SERVICE_ACCOUNT_KEY" | base64 -d > gcp-sa.json
 else
     echo "Secret já está no formato correto, salvando diretamente..."
-    echo "$GCP_SERVICE_ACCOUNT_KEY" > gcp-sa.json
+    echo "$SERVICE_ACCOUNT_KEY" > gcp-sa.json
 fi
 
-# Autenticar o gcloud
-gcloud auth activate-service-account --key-file=gcp-sa.json
-# Configurar o Docker para autenticar com o GCR
-gcloud auth configure-docker us-docker.pkg.dev
+if [$DEPLOY_PROVIDER == "GCP"]; then
+  # Autenticar o gcloud
+  gcloud auth activate-service-account --key-file=gcp-sa.json
+  # Configurar o Docker para autenticar com o GCR
+  gcloud auth configure-docker us-docker.pkg.dev
+elif [$DEPLOY_PROVIDER == "AWS"]; then
+  # Autenticar o AWS CLI
+  aws configure set aws_access_key_id "$AWS_ACCESS_KEY_ID"
+  aws configure set aws_secret_access_key "$AWS_SECRET_ACCESS_KEY"
+  aws configure set default.region "$AWS_REGION"
+
+  # Autenticar o Docker com o ECR
+  aws ecr get-login-password --region "$AWS_REGION" | docker login --username AWS --password-stdin "$REPOSITORY_URI_BRANCH"
+fi
 
 # Verificar se a branch é master ou main
 if [[ "$GITHUB_REF_NAME" == "master" || "$GITHUB_REF_NAME" == "main" ]]; then
@@ -102,17 +135,31 @@ if [[ "$GITHUB_REF_NAME" == "master" || "$GITHUB_REF_NAME" == "main" ]]; then
       echo "Build e push concluídos para frontend em $GITHUB_REF_NAME."
 
   else
-      LATEST_IMAGE_LINE=$(gcloud artifacts docker images list --include-tags "$REPOSITORY_URI_PRD" \
-          --sort-by=~UPDATE_TIME \
-          --limit=1 \
-          --quiet \
-          | tail -n 1)
+      
+      if [ "$DEPLOY_PROVIDER" == "GCP" ]; then
+        LATEST_IMAGE_LINE=$(gcloud artifacts docker images list --include-tags "$REPOSITORY_URI_PRD" \
+            --sort-by=~UPDATE_TIME \
+            --limit=1 \
+            --quiet \
+            | tail -n 1)
 
-      if [[ -z "$LATEST_IMAGE_LINE" ]]; then
-          echo "Erro Crítico: Nenhuma linha de dados retornada por gcloud."
-          exit 1
-      fi
+        if [[ -z "$LATEST_IMAGE_LINE" ]]; then
+            echo "Erro Crítico: Nenhuma linha de dados retornada por gcloud."
+            exit 1
+        fi
+      elif  [ "$DEPLOY_PROVIDER" == "AWS" ]; then
+          LATEST_IMAGE_LINE=$(aws ecr describe-images \
+            --repository-name "$REPOSITORY_NAME" \
+            --region "$AWS_REGION" \
+            --query 'sort_by(imageDetails,& imagePushedAt)[-1]' \
+            --output json)
 
+        if [[ -z "$LATEST_IMAGE_LINE" ]]; then
+            echo "Erro Crítico: Nenhuma linha de dados retornada por gcloud."
+            exit 1
+        fi
+        fi
+      
       IMAGE_DIGEST=$(echo "$LATEST_IMAGE_LINE" | awk '{print $2}')
       IMAGE_TAG=$(echo "$LATEST_IMAGE_LINE" | awk '{print $3}')
 
@@ -182,4 +229,14 @@ else
   echo "Outputs definidos"
 fi
 
-rm -f gcp-sa.json
+if [ "$DEPLOY_PROVIDER" == "GCP" ]; then
+    echo "Removendo arquivo de chave do GCP..."
+    rm -f gcp-sa.json
+elif [ "$DEPLOY_PROVIDER" == "AWS" ]; then
+    echo "Removendo credenciais do AWS..."
+    unset AWS_ACCESS_KEY_ID
+    unset AWS_SECRET_ACCESS_KEY
+fi
+
+echo "Script de build e push concluído com sucesso."
+echo "IMAGE_TAG=$IMAGE_TAG" >> "$GITHUB_OUTPUT"
